Make it possible to actually follow the instructions and run the build script (even if it presently produces no useful output):
1. Update/correct/simplify the instructions.

2. Update/correct/simplify the default config, and make the code
which uses it a little less clumsy. This is not complete, but is at
least a start.

3. Update the kf5.py code to be less - ahem - idiosyncratic. The
changes *look* huge, but a diff ignoring whitespace will reveal the
reality that the goal is simply to make setting up the the config
from step 2 actually work in a roughly sane fashion.